3.6 terminado

+++ SegundoTrimestre/spring_boot_web_mvc_jdbc_ventas-master/src/main/java/org/iesbelen/modelo/Pedido.java
@@ -12,6 +12,6 @@ public class Pedido {
     private Integer id;
     private double total;
     private LocalDate fecha;
-    private Integer id_cliente;
+    private Long id_cliente;
     private Integer id_comercial;
 }
